backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
PyPDF2==3.0.1
python-dotenv==1.0.1
google-generativeai==0.8.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.36
transformers==4.47.1
torch==2.5.1
